feat(widget-v2)/balance loading state
