+++ packages/widget-v2/src/icons/SpinnerIcon.tsx
@@ -1,0 +1,26 @@
+import type { ComponentProps } from "react";
+
+export const SpinnerIcon = (props: ComponentProps<"svg">) => {
+  return (
+    <svg
+      fill="none"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <circle
+        cx="12"
+        cy="12"
+        opacity=".25"
+        r="10"
+        stroke="currentColor"
+        strokeWidth="4"
+      />
+      <path
+        d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+        fill="currentColor"
+        opacity=".75"
+      />
+    </svg>
+  );
+};
